refactor(web): add barrel index.ts to every component folder
Each feature folder under components/ now exposes an index.ts barrel so
consumers import from "@/components/<folder>" instead of deep file paths:

- Added barrels: auth, combo, dashboard, keys, layout, logs, playground,
  providers, settings, tokenSaver (analytics + skeletons already had one)
- Rewired all alias imports to the folder root
- Added named exports for previously-default-only components
  (KeyMetrics, ComboForm) so barrel consumers use named imports
- Merged duplicate imports from the same source into single statements
- ui/ intentionally keeps per-file imports (shadcn convention)

Build (tsc + vite) and lint pass.
