+++ apps/web/src/components/auth/index.ts
@@ -1,0 +1,1 @@
+export { AdminAuthGate } from "./AdminAuthGate";
