+++ apps/web/src/components/combo/index.ts
@@ -1,0 +1,5 @@
+export { ComboArchitecture } from "./combo.architecture";
+export { default as ComboModelPickerModal, getModelCapabilities, formatModelDisplayName, type ComboModelItem } from "./combo.dialog";
+export { default as ComboForm } from "./combo.form";
+export { ComboHeader } from "./combo.header";
+export { ComboList } from "./combo.list";
