+++ apps/web/src/components/keys/keys.section.tsx
@@ -87,3 +87,5 @@ export default function KeyMetrics({
         </section>
     );
 }
+
+export { KeyMetrics };
