+++ apps/web/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { AppSidebar } from "./AppSidebar";
+export { Topbar } from "./Topbar";
