+++ apps/web/src/components/logs/index.ts
@@ -1,0 +1,2 @@
+export { LogDetailSheet } from "./logs.detail-sheet";
+export { LogTable } from "./logs.table";
