+++ apps/web/src/components/providers/providers.connect-oauth-modal.tsx
@@ -22,7 +22,7 @@ import {
     DialogTitle
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { ProviderIcon } from "@/components/providers/providers.icon";
+import { ProviderIcon } from "@/components/providers";
 
 interface ConnectOAuthModalProps {
     provider: ProviderDefinition | null;
