+++ apps/web/src/components/settings/index.ts
@@ -1,0 +1,8 @@
+export { AppearanceSettings } from "./settings.appearance";
+export { DataSettings } from "./settings.data";
+export { GatewaySettings } from "./settings.gateway";
+export { LoggingSettings } from "./settings.logging";
+export { PlaygroundSettings } from "./settings.playground";
+export { SecuritySettings } from "./settings.security";
+export { SystemSettings } from "./settings.system";
+export { SettingsSection, SettingsRow, SegmentedControl, ValueBadge } from "./settings.ui";
