+++ apps/web/src/routes/logs.tsx
@@ -7,8 +7,7 @@ import type { RequestLogEntry } from "@srouter/types";
 import type { ListResponse } from "@/lib/types";
 import { LogsSkeleton } from "@/components/skeletons";
 import { useLogs } from "@/hooks/useLogs";
-import { LogTable } from "@/components/logs/logs.table";
-import { LogDetailSheet } from "@/components/logs/logs.detail-sheet";
+import { LogDetailSheet, LogTable } from "@/components/logs";
 
 export const Route = createFileRoute("/logs")({
     staticData: { title: "Logs" },
