+++ apps/web/src/routes/providers/index.tsx
@@ -3,9 +3,7 @@ import { createFileRoute } from "@tanstack/react-router";
 import { RefreshCw, TriangleAlert } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { useCatalog } from "@/hooks/useCatalog";
-import { Catalog } from "@/components/providers/providers.catalog";
-import { CatalogToolbar } from "@/components/providers/providers.catalog-toolbar";
-import { CustomProviderDialog } from "@/components/providers/providers.custom-provider-dialog";
+import { Catalog, CatalogToolbar, CustomProviderDialog } from "@/components/providers";
 import { ProvidersSkeleton } from "@/components/skeletons";
 
 export const Route = createFileRoute("/providers/")({
