+++ apps/web/src/routes/quota.tsx
@@ -16,7 +16,7 @@ import {
 import { toast } from "sonner";
 import { QuotaSkeleton } from "@/components/skeletons";
 import { Button, buttonVariants } from "@/components/ui/button";
-import { ProviderIcon } from "@/components/providers/providers.icon";
+import { ProviderIcon } from "@/components/providers";
 import { cn } from "@/lib/utils";
 import { useQuota } from "@/hooks/useQuota";
 import type { LiveModelQuotaItem, ProviderUsageMetric } from "@srouter/types";
