+++ apps/web/src/routes/token-saver.tsx
@@ -1,8 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { useTokenSaver } from "@/hooks/useTokenSaver";
-import { TokenSaverHeader } from "@/components/tokenSaver/tokenSaver.header";
-import { ToolCompressionCard } from "@/components/tokenSaver/tokenSaver.tool-compression-card";
-import { PromptOptimizerCard } from "@/components/tokenSaver/tokenSaver.prompt-optimizer-card";
+import { PromptOptimizerCard, TokenSaverHeader, ToolCompressionCard } from "@/components/tokenSaver";
 
 export const Route = createFileRoute("/token-saver")({
     staticData: { title: "Token Saver" },
